fix: bump opentelemetry-proto to 0.32 to patch opentelemetry_sdk CVE
opentelemetry_sdk <= 0.32.0 (pulled in transitively via
opentelemetry-proto's `logs` feature) has GHSA-w9wp-h8wv-79jx /
CVE-2026-48504: BaggagePropagator::extract_with_context doesn't
enforce W3C baggage header size limits before parsing, allowing DoS
via oversized headers. Cortex's OTLP receiver only deserializes
protobuf message bodies via prost and never calls the baggage
propagator, so the vulnerable path isn't reachable here, but the dep
should still carry the fix.

opentelemetry-proto 0.32 also adds a new AnyValue::StringValueStrindex
wire variant (string-table-indexed encoding); handle it the same way
as the other composite AnyValue kinds by rendering a placeholder
instead of expanding inline.

Splitting otlp.rs was needed to stay under the 500-line module-size
gate once the new match arms landed: AnyValue/LogBatchEntry
conversion moves to otlp/entries.rs and the bearer-token auth gate
moves to otlp/auth.rs, following the existing foo.rs-sibling-to-foo/
convention (src/mcp.rs, src/db.rs). otlp.rs keeps the router, state,
and HTTP handlers; the public surface (OtlpState, OtlpCounters,
router, OTLP_BODY_LIMIT_BYTES) is unchanged.
