+++ src/otlp.rs
@@ -5,40 +5,39 @@
 //! Mounted on the same axum server as MCP. Body limit: 4 MiB. Optional Bearer
 //! auth via the same `CORTEX_TOKEN` as MCP (`CORTEX_API_TOKEN` is
 //! accepted as a deprecated alias).
+//!
+//! Request → response wiring lives here; `AnyValue`/`LogBatchEntry`
+//! conversion is in [`entries`] and the bearer-token gate is in [`auth`].
 
-use std::collections::HashMap;
 use std::net::SocketAddr;
+use std::sync::Arc;
 use std::sync::atomic::{AtomicU64, Ordering};
-use std::sync::{Arc, LazyLock, Mutex};
-use std::time::{Duration, Instant};
 
 use crate::mcp::AuthPolicy;
 use axum::{
     Router,
     extract::{ConnectInfo, State},
-    http::{
-        HeaderMap, HeaderValue, StatusCode,
-        header::{AUTHORIZATION, RETRY_AFTER, USER_AGENT},
-    },
+    http::{HeaderMap, HeaderValue, StatusCode, header::RETRY_AFTER},
     middleware::{Next, from_fn},
     response::{IntoResponse, Json},
     routing::post,
 };
 use bytes::Bytes;
-use opentelemetry_proto::tonic::{
-    collector::logs::v1::ExportLogsServiceRequest,
-    common::v1::{AnyValue, any_value::Value as AnyValueKind},
-};
+use opentelemetry_proto::tonic::collector::logs::v1::ExportLogsServiceRequest;
 use prost::Message;
 use serde_json::json;
-use sha2::{Digest, Sha256};
 use tower_http::limit::RequestBodyLimitLayer;
 
-use crate::db::LogBatchEntry;
-use crate::enrich::{SourceKind, stamp_source_kind};
 use crate::ingest::IngestTx;
-use crate::ingest_metadata::{attrs_to_metadata_object, bounded_metadata_json};
-use lab_auth::middleware::{parse_bearer_token, tokens_equal};
+
+mod auth;
+mod entries;
+
+use auth::{
+    is_authorized, otlp_auth_policy_label, should_warn_unauthorized, unauthorized,
+    unauthorized_diagnostics,
+};
+use entries::build_entries;
 
 /// Per-request body cap. Matches the OpenTelemetry Collector default for
 /// HTTP receivers. Larger payloads receive 413 + `Retry-After: 86400`.
@@ -276,340 +275,6 @@ async fn traces_handler(
             "error": "traces_not_supported",
             "message": "OTLP traces deferred. Use /v1/logs only."
         })),
-    )
-        .into_response()
-}
-
-/// Walk the OTLP request and produce one `LogBatchEntry` per `LogRecord`.
-fn build_entries(req: &ExportLogsServiceRequest, peer: SocketAddr) -> Vec<LogBatchEntry> {
-    let received_iso = chrono::Utc::now()
-        .format("%Y-%m-%dT%H:%M:%S%.3fZ")
-        .to_string();
-    let source_ip = peer.to_string();
-    let peer_ip = peer.ip().to_string();
-
-    let mut out = Vec::new();
-    for resource_logs in &req.resource_logs {
-        let resource_attrs: HashMap<&str, &AnyValue> = resource_logs
-            .resource
-            .as_ref()
-            .map(|r| {
-                r.attributes
-                    .iter()
-                    .filter_map(|kv| kv.value.as_ref().map(|v| (kv.key.as_str(), v)))
-                    .collect()
-            })
-            .unwrap_or_default();
-        let hostname = resource_attrs
-            .get("host.name")
-            .and_then(|v| any_value_to_string(v))
-            .unwrap_or_default();
-        let service_name = resource_attrs
-            .get("service.name")
-            .and_then(|v| any_value_to_string(v));
-        let service_version = resource_attrs
-            .get("service.version")
-            .and_then(|v| any_value_to_string(v));
-
-        for scope_logs in &resource_logs.scope_logs {
-            for log in &scope_logs.log_records {
-                let log_attrs: HashMap<&str, &AnyValue> = log
-                    .attributes
-                    .iter()
-                    .filter_map(|kv| kv.value.as_ref().map(|v| (kv.key.as_str(), v)))
-                    .collect();
-
-                let ai_session_id = log_attrs
-                    .get("session.id")
-                    .or_else(|| log_attrs.get("session_id"))
-                    .or_else(|| resource_attrs.get("session.id"))
-                    .or_else(|| resource_attrs.get("session_id"))
-                    .and_then(|v| any_value_to_string(v))
-                    .filter(|value| value.len() <= 128);
-
-                let ai_project = log_attrs
-                    .get("project.path")
-                    .or_else(|| log_attrs.get("codebase.root_path"))
-                    .or_else(|| log_attrs.get("session.cwd"))
-                    .or_else(|| resource_attrs.get("project.path"))
-                    .or_else(|| resource_attrs.get("codebase.root_path"))
-                    .or_else(|| resource_attrs.get("session.cwd"))
-                    .and_then(|v| any_value_to_string(v))
-                    .filter(|value| value.len() <= 512);
-
-                let timestamp = format_otlp_timestamp(log.time_unix_nano)
-                    .unwrap_or_else(|| received_iso.clone());
-                let severity = severity_from_number(log.severity_number).to_string();
-                let message = log
-                    .body
-                    .as_ref()
-                    .and_then(any_value_to_string)
-                    .unwrap_or_default();
-                let metadata_json = bounded_metadata_json(serde_json::json!({
-                    "source_type": "otlp",
-                    "peer_ip": peer_ip,
-                    "peer_port": peer.port(),
-                    "host_name": hostname,
-                    "service_name": service_name,
-                    "service_version": service_version,
-                    "severity_number": log.severity_number,
-                    "severity_text": log.severity_text,
-                    "trace_id": hex_bytes(&log.trace_id),
-                    "span_id": hex_bytes(&log.span_id),
-                    "flags": log.flags,
-                    "event_name": log.event_name,
-                    "resource_attributes": attrs_to_json(&resource_attrs),
-                    "log_attributes": attrs_to_json(&log_attrs),
-                }));
-                let mut entry = LogBatchEntry {
-                    timestamp,
-                    hostname: hostname.clone(),
-                    facility: Some("otlp".to_string()),
-                    severity,
-                    app_name: service_name.clone(),
-                    process_id: None,
-                    message,
-                    raw: metadata_json.clone(),
-                    source_ip: source_ip.clone(),
-                    docker_checkpoint: None,
-                    ai_tool: extract_ai_tool(&log_attrs, &resource_attrs),
-                    ai_project,
-                    ai_session_id,
-                    ai_transcript_path: None,
-                    metadata_json: Some(metadata_json),
-                    http_status: None,
-                    auth_outcome: None,
-                    dns_blocked: None,
-                    event_action: None,
-                    parse_error: None,
-                };
-                stamp_source_kind(&mut entry, SourceKind::Otlp);
-                out.push(entry);
-            }
-        }
-    }
-    out
-}
-
-fn attrs_to_json(attrs: &HashMap<&str, &AnyValue>) -> serde_json::Value {
-    attrs_to_metadata_object(
-        attrs
-            .iter()
-            .map(|(key, value)| (*key, any_value_to_json(value))),
-    )
-}
-
-fn any_value_to_json(v: &AnyValue) -> serde_json::Value {
-    match v.value.as_ref() {
-        Some(AnyValueKind::StringValue(s)) => serde_json::Value::String(s.clone()),
-        Some(AnyValueKind::BoolValue(b)) => serde_json::Value::Bool(*b),
-        Some(AnyValueKind::IntValue(i)) => serde_json::Value::Number((*i).into()),
-        Some(AnyValueKind::DoubleValue(f)) => serde_json::Number::from_f64(*f)
-            .map(serde_json::Value::Number)
-            .unwrap_or(serde_json::Value::Null),
-        Some(AnyValueKind::BytesValue(b)) => serde_json::json!({"bytes_len": b.len()}),
-        Some(AnyValueKind::ArrayValue(arr)) => serde_json::json!({"array_len": arr.values.len()}),
-        Some(AnyValueKind::KvlistValue(kv)) => serde_json::json!({"kvlist_len": kv.values.len()}),
-        None => serde_json::Value::Null,
-    }
-}
-
-fn hex_bytes(bytes: &[u8]) -> Option<String> {
-    if bytes.is_empty() {
-        return None;
-    }
-    Some(bytes.iter().map(|byte| format!("{byte:02x}")).collect())
-}
-
-fn extract_ai_tool(
-    log_attrs: &HashMap<&str, &AnyValue>,
-    resource_attrs: &HashMap<&str, &AnyValue>,
-) -> Option<String> {
-    let raw = log_attrs
-        .get("ai.tool")
-        .or_else(|| log_attrs.get("ai_tool"))
-        .or_else(|| resource_attrs.get("ai.tool"))
-        .or_else(|| resource_attrs.get("ai_tool"))
-        .and_then(|v| any_value_to_string(v))?;
-    if raw.len() > 64 {
-        return None;
-    }
-    match raw.to_ascii_lowercase().as_str() {
-        "claude" | "codex" | "gemini" => Some(raw.to_ascii_lowercase()),
-        _ => None,
-    }
-}
-
-fn format_otlp_timestamp(time_unix_nano: u64) -> Option<String> {
-    if time_unix_nano == 0 {
-        return None;
-    }
-    let secs = (time_unix_nano / 1_000_000_000) as i64;
-    let nanos = (time_unix_nano % 1_000_000_000) as u32;
-    chrono::DateTime::<chrono::Utc>::from_timestamp(secs, nanos)
-        .map(|dt| dt.format("%Y-%m-%dT%H:%M:%S%.3fZ").to_string())
-}
-
-/// OTLP `SeverityNumber` (0–24) → syslog severity string.
-///
-/// 0 (UNSPECIFIED) and any unrecognised value fall through to `info` rather
-/// than dropping the record.
-fn severity_from_number(n: i32) -> &'static str {
-    match n {
-        1..=8 => "debug", // OTLP TRACE (1..=4) and DEBUG (5..=8) both map here
-        9..=12 => "info",
-        13..=16 => "warning",
-        17..=20 => "err",
-        21..=24 => "crit",
-        _ => "info", // 0=UNSPECIFIED and out-of-range fall back to info
-    }
-}
-
-/// Stringify any `AnyValue` variant for storage in `message` / attribute fields.
-/// Composite types render as a placeholder rather than expanding inline.
-fn any_value_to_string(v: &AnyValue) -> Option<String> {
-    match v.value.as_ref()? {
-        AnyValueKind::StringValue(s) => Some(s.clone()),
-        AnyValueKind::BoolValue(b) => Some(b.to_string()),
-        AnyValueKind::IntValue(i) => Some(i.to_string()),
-        AnyValueKind::DoubleValue(f) => Some(f.to_string()),
-        AnyValueKind::BytesValue(b) => Some(format!("[{} bytes]", b.len())),
-        AnyValueKind::ArrayValue(arr) => Some(format!("[array len={}]", arr.values.len())),
-        AnyValueKind::KvlistValue(kv) => Some(format!("[kvlist len={}]", kv.values.len())),
-    }
-}
-
-fn is_authorized(state: &OtlpState, headers: &HeaderMap) -> bool {
-    // No-auth policies: loopback bind or upstream gateway is the trust boundary.
-    if matches!(
-        state.auth_policy,
-        AuthPolicy::LoopbackDev | AuthPolicy::TrustedGatewayUnscoped
-    ) {
-        return true;
-    }
-    // Mounted auth: require the static bearer token. If none is configured
-    // (OAuth-only deployment), OTLP is denied — there is no OAuth flow for
-    // machine-to-machine OTLP exporters.
-    let Some(expected) = state.api_token.as_deref() else {
-        return false;
-    };
-    let Some(auth) = headers.get(AUTHORIZATION).and_then(|v| v.to_str().ok()) else {
-        return false;
-    };
-    parse_bearer_token(auth).is_some_and(|tok| tokens_equal(&tok, expected))
-}
-
-#[derive(Debug, PartialEq, Eq)]
-struct UnauthorizedDiagnostics {
-    has_auth: bool,
-    auth_scheme: String,
-    bearer_sha256_12: String,
-    user_agent: String,
-}
-
-static UNAUTHORIZED_WARNINGS: LazyLock<Mutex<HashMap<String, Instant>>> =
-    LazyLock::new(|| Mutex::new(HashMap::new()));
-
-const UNAUTHORIZED_WARNING_INTERVAL: Duration = Duration::from_secs(60);
-const UNAUTHORIZED_WARNING_MAX_KEYS: usize = 1024;
-const MAX_DIAGNOSTIC_FIELD_LEN: usize = 128;
-
-fn should_warn_unauthorized(peer: &SocketAddr, diagnostics: &UnauthorizedDiagnostics) -> bool {
-    let key = unauthorized_warning_key(peer, diagnostics);
-    let now = Instant::now();
-    let Ok(mut warnings) = UNAUTHORIZED_WARNINGS.lock() else {
-        return true;
-    };
-    record_unauthorized_warning(
-        &mut warnings,
-        key,
-        now,
-        UNAUTHORIZED_WARNING_INTERVAL,
-        UNAUTHORIZED_WARNING_MAX_KEYS,
-    )
-}
-
-fn unauthorized_warning_key(peer: &SocketAddr, diagnostics: &UnauthorizedDiagnostics) -> String {
-    format!(
-        "{}|{}|{}|{}",
-        peer.ip(),
-        diagnostics.auth_scheme,
-        diagnostics.bearer_sha256_12,
-        diagnostics.user_agent
-    )
-}
-
-fn record_unauthorized_warning(
-    warnings: &mut HashMap<String, Instant>,
-    key: String,
-    now: Instant,
-    interval: Duration,
-    max_keys: usize,
-) -> bool {
-    match warnings.get(&key).copied() {
-        Some(last) if now.duration_since(last) < interval => false,
-        _ => {
-            if warnings.len() >= max_keys {
-                warnings.retain(|_, last| now.duration_since(*last) < interval);
-            }
-            if !warnings.contains_key(&key) && warnings.len() >= max_keys {
-                return false;
-            }
-            warnings.insert(key, now);
-            true
-        }
-    }
-}
-
-fn unauthorized_diagnostics(headers: &HeaderMap) -> UnauthorizedDiagnostics {
-    let auth = headers.get(AUTHORIZATION).and_then(|v| v.to_str().ok());
-    let bearer = auth.and_then(parse_bearer_token);
-    UnauthorizedDiagnostics {
-        has_auth: auth.is_some(),
-        auth_scheme: auth_scheme(auth),
-        bearer_sha256_12: bearer
-            .as_deref()
-            .map(sha256_12)
-            .unwrap_or_else(|| "none".to_string()),
-        user_agent: headers
-            .get(USER_AGENT)
-            .and_then(|v| v.to_str().ok())
-            .filter(|value| !value.trim().is_empty())
-            .map(truncate_diagnostic_field)
-            .unwrap_or_else(|| "unknown".to_string()),
-    }
-}
-
-fn auth_scheme(auth: Option<&str>) -> String {
-    auth.and_then(|value| value.split_ascii_whitespace().next())
-        .filter(|scheme| !scheme.is_empty())
-        .unwrap_or("none")
-        .to_ascii_lowercase()
-}
-
-fn sha256_12(value: &str) -> String {
-    let mut hasher = Sha256::new();
-    hasher.update(value.as_bytes());
-    let digest = hasher.finalize();
-    format!("{digest:x}")[..12].to_string()
-}
-
-fn truncate_diagnostic_field(value: &str) -> String {
-    value.chars().take(MAX_DIAGNOSTIC_FIELD_LEN).collect()
-}
-
-fn otlp_auth_policy_label(policy: &AuthPolicy) -> &'static str {
-    match policy {
-        AuthPolicy::LoopbackDev => "loopback_dev",
-        AuthPolicy::TrustedGatewayUnscoped => "trusted_gateway",
-        AuthPolicy::Mounted { .. } => "mounted",
-    }
-}
-
-fn unauthorized() -> axum::response::Response {
-    (
-        StatusCode::UNAUTHORIZED,
-        Json(json!({"error": "unauthorized"})),
     )
         .into_response()
 }
